Exercicios da aula 2 e aula 3 sobre Funções e Objetos

+++ src/e2.kt
@@ -3,6 +3,6 @@ fun main(){
     val n1 = readLine()!!
     print("Digite outro numero: ")
     val n2 = readLine()!!
-    val soma  = n1 + n2
+    val soma  = n1.toInt() + n2.toInt()
     println("A soma dos números é $soma")
 }
